+++ test/index.test.js
@@ -15,7 +15,12 @@ describe('Amplitude', function() {
     trackReferrer: false,
     batchEvents: false,
     eventUploadThreshold: 30,
-    eventUploadPeriodMillis: 30000
+    eventUploadPeriodMillis: 30000,
+    forceHttps: false,
+    trackGclid: false,
+    saveParamsReferrerOncePerSession: true,
+    deviceIdFromUrlParam: false,
+    mapQueryParams: {}
   };
 
   beforeEach(function() {
@@ -43,7 +48,11 @@ describe('Amplitude', function() {
       .option('trackReferrer', false)
       .option('batchEvents', false)
       .option('eventUploadThreshold', 30)
-      .option('eventUploadPeriodMillis', 30000));
+      .option('eventUploadPeriodMillis', 30000)
+      .option('forceHttps', false)
+      .option('trackGclid', false)
+      .option('saveParamsReferrerOncePerSession', true)
+      .option('deviceIdFromUrlParam', false));
   });
 
   describe('before loading', function() {
@@ -89,6 +98,10 @@ describe('Amplitude', function() {
       analytics.assert(window.amplitude.options.batchEvents === options.batchEvents);
       analytics.assert(window.amplitude.options.eventUploadThreshold === options.eventUploadThreshold);
       analytics.assert(window.amplitude.options.eventUploadPeriodMillis === options.eventUploadPeriodMillis);
+      analytics.assert(window.amplitude.options.forceHttps === options.forceHttps);
+      analytics.assert(window.amplitude.options.includeGclid === options.trackGclid);
+      analytics.assert(window.amplitude.options.saveParamsReferrerOncePerSession === options.saveParamsReferrerOncePerSession);
+      analytics.assert(window.amplitude.options.deviceIdFromUrlParam === options.deviceIdFromUrlParam);
     });
 
     it('should set api key', function() {
@@ -130,6 +143,7 @@ describe('Amplitude', function() {
     describe('#page', function() {
       beforeEach(function() {
         analytics.stub(window.amplitude, 'logEvent');
+        analytics.stub(window.amplitude, 'setUserProperties');
       });
 
       it('should not track unnamed pages by default', function() {
@@ -164,12 +178,34 @@ describe('Amplitude', function() {
         analytics.page('Category', 'Name');
         analytics.didNotCall(window.amplitude.logEvent);
       });
+
+      it('should map query params to custom property as user properties', function() {
+        amplitude.options.trackAllPages = true;
+        amplitude.options.mapQueryParams = { customProp: 'user_properties' };
+        analytics.page({}, { page: { search: '?suh=dude' } });
+        analytics.called(window.amplitude.setUserProperties, { customProp: '?suh=dude' });
+      });
+
+      it('should map query params to custom property as event properties', function() {
+        amplitude.options.trackAllPages = true;
+        amplitude.options.mapQueryParams = { params: 'event_properties' };
+        analytics.page({ referrer: document.referrer }, { page: { search: '?suh=dude' } });
+        analytics.called(window.amplitude.logEvent, 'Loaded a Page', {
+          params: '?suh=dude',
+          path: '/context.html',
+          referrer: document.referrer,
+          search: '', // in practice this would also be set to the query param but limitation of test prevents this from being set
+          title: '',
+          url: 'http://localhost:9876/context.html'
+        });
+      });
     });
 
     describe('#identify', function() {
       beforeEach(function() {
         analytics.stub(window.amplitude, 'setUserId');
         analytics.stub(window.amplitude, 'setUserProperties');
+        analytics.stub(window.amplitude, 'setGroup');
       });
 
       it('should send an id', function() {
@@ -187,13 +223,27 @@ describe('Amplitude', function() {
         analytics.called(window.amplitude.setUserId, 'id');
         analytics.called(window.amplitude.setUserProperties, { id: 'id', trait: true });
       });
+
+      it('should send query params under custom trait if set', function() {
+        amplitude.options.mapQueryParams = { ham: 'user_properties' };
+        analytics.identify('id', { trait: true }, { page: { search: '?foo=bar' } });
+        analytics.called(window.amplitude.setUserId, 'id');
+        analytics.called(window.amplitude.setUserProperties, { id: 'id', trait: true, ham: '?foo=bar' });
+      });
+
+      it('should set user groups if integration option `groups` is present', function() {
+        analytics.identify('id', {}, { integrations: { Amplitude: { groups: { foo: 'bar' } } } });
+        analytics.called(window.amplitude.setGroup, 'foo', 'bar');
+      });
     });
 
     describe('#track', function() {
       beforeEach(function() {
         analytics.stub(window.amplitude, 'logEvent');
+        analytics.stub(window.amplitude, 'setUserProperties');
         analytics.stub(window.amplitude, 'logRevenue');
         analytics.stub(window.amplitude, 'logRevenueV2');
+        analytics.stub(window.amplitude, 'logEventWithGroups');
       });
 
       it('should send an event', function() {
@@ -262,6 +312,23 @@ describe('Amplitude', function() {
         analytics.called(window.amplitude.logEvent);
         analytics.didNotCall(window.amplitude.logRevenue);
         analytics.didNotCall(window.amplitude.logRevenueV2);
+      });
+
+      it('should send a query params under custom prop as user properties', function() {
+        amplitude.options.mapQueryParams = { ham: 'user_properties' };
+        analytics.track('event', { foo: 'bar' }, { page: { search: '?foo=bar' } });
+        analytics.called(window.amplitude.setUserProperties, { ham: '?foo=bar' });
+      });
+
+      it('should send a query params under custom prop as user or event properties', function() {
+        amplitude.options.mapQueryParams = { ham: 'event_properties' };
+        analytics.track('event', { foo: 'bar' }, { page: { search: '?foo=bar' } });
+        analytics.called(window.amplitude.logEvent, 'event', { foo: 'bar', ham: '?foo=bar' });
+      });
+
+      it('should send an event with groups if `groups` is an integration specific option', function() {
+        analytics.track('event', { foo: 'bar' }, { integrations: { Amplitude: { groups: { sports: 'basketball' } } } });
+        analytics.called(window.amplitude.logEventWithGroups, 'event', { foo: 'bar' }, { sports: 'basketball' });
       });
     });